+++ lotti/lib/pages/settings/settings_card.dart
@@ -16,7 +16,7 @@ class SettingsCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
-      color: AppColors.headerBgColor,
+      color: AppColors.entryCardColor,
       elevation: 8.0,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8.0),