+++ lotti/lib/widgets/journal/editor/editor_wrapper_widget.dart
@@ -8,6 +8,7 @@ import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/editor_state_service.dart';
+import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/journal/editor/editor_tools.dart';
 import 'package:lotti/widgets/journal/editor/editor_widget.dart';
 import 'package:path_provider/path_provider.dart';
@@ -83,7 +84,10 @@ class _EditorWrapperWidgetState extends State<EditorWrapperWidget> {
 
         void saveText() {
           _editorStateService.saveState(widget.itemId, _controller);
-          _focusNode.unfocus();
+
+          if (isMobile) {
+            _focusNode.unfocus();
+          }
         }
 
         return item.maybeMap(