+++ lotti/lib/widgets/journal/entry_details_widget.dart
@@ -13,6 +13,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/editor_state_service.dart';
 import 'package:lotti/theme.dart';
+import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/audio/audio_player.dart';
 import 'package:lotti/widgets/journal/editor/editor_tools.dart';
 import 'package:lotti/widgets/journal/editor/editor_widget.dart';
@@ -104,11 +105,14 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
 
         void saveText() {
           _editorStateService.saveState(widget.itemId, _controller);
-          _focusNode.unfocus();
+
+          if (isMobile) {
+            _focusNode.unfocus();
+          }
         }
 
         return Container(
-          margin: const EdgeInsets.symmetric(horizontal: 4, vertical: 6),
+          margin: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
           decoration: BoxDecoration(
             boxShadow: [
               BoxShadow(
@@ -122,7 +126,7 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
           child: ClipRRect(
             borderRadius: BorderRadius.circular(8.0),
             child: Container(
-              color: AppColors.headerBgColor,
+              color: AppColors.entryCardColor,
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [