+++ lotti/CHANGELOG.md
@@ -5,15 +5,31 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.10] - 2022-04-26
 ### Added:
-- text color in sync settings
-- unfocus on save
-- entry styling
+- Sync settings: hide sensitive info
+- Sync settings: select IMAP folder
+- Styling: shadow on navigation bar
+- Sync assistant
+
+### Changed:  
+- Task form styling
+- Unfocus on save only on mobile
+- Sync IMAP messages marked seen
+- Sync IMAP messages in lotti_sync folder
+- Styling: entry card color
+
+## [0.7.9] - 2022-04-25
+### Changed:
+- Text color in sync settings
+- Unfocus on save
+- Entry styling
 
 ## [0.7.8] - 2022-04-24
 ### Added:
-- new color scheme
-- text editor in slideshow for faster import
+- New color scheme
+- Text editor in slideshow for faster import
 
 ## [0.7.7] - 2022-04-24
 ### Added: